refactor: 파일 저장 관련 Filedata로변경

+++ domain/src/main/kotlin/org/chewing/v1/implementation/media/FileProcessor.kt
@@ -1,31 +1,38 @@
 package org.chewing.v1.implementation.media
 
+import org.chewing.v1.model.media.FileCategory
+import org.chewing.v1.model.media.FileData
 import org.chewing.v1.model.media.Media
 import org.springframework.stereotype.Component
 import java.io.File
+
 @Component
 class FileProcessor(
     private val fileProvider: FileProvider,
     private val fileGenerator: FileGenerator,
     private val fileValidator: FileValidator
 ) {
-    fun processNewFiles(userId: String, files: List<File>): List<Media> {
+    fun processNewFiles(userId: String, files: List<FileData>, category: FileCategory): List<Media> {
         fileValidator.validateFilesNameCorrect(files)
-        val mediaWithFiles = fileGenerator.generateFeedMedias(files, userId)
+        val mediaWithFiles = fileGenerator.generateMedias(files, userId, category)
         mediaWithFiles.forEach {
             val (file, media) = it
             fileProvider.appendFile(file, media)
         }
         return mediaWithFiles.map { it.second }
     }
-    fun processNewFile(userId: String, file: File): Media {
+
+    fun processNewFile(userId: String, file: FileData, category: FileCategory): Media {
         fileValidator.validateFileNameCorrect(file)
-        val media = fileGenerator.generateFeedMedia(file, userId)
+        val media = fileGenerator.generateMedia(file, userId, category)
         fileProvider.appendFile(file, media)
         return media
     }
+
     fun processOldFile(media: Media) {
-        fileProvider.removeFile(media)
+        if (!media.isBasic) {
+            fileProvider.removeFile(media)
+        }
     }
 
     fun processOldFiles(medias: List<Media>) {
